Refactoring, and touch-ups
seperated purchasing from writeoffs, product addition, stocking and product deletion to an other page. new page requires a password to access. inventory cannot dip below 0, need to find a way to send a notification when the program tries to though to let user know that stock is incorrect for that product, code is set up to send the message.